move exec_git_command method to Utils::Git module
