+++ lib/datadog/ci/ext/environment/providers/local_git.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
-require "open3"
-
 require_relative "base"
-require_relative "../../git"
+require_relative "../../../utils/git"
 
 module Datadog
   module CI
@@ -13,7 +11,7 @@ module Datadog
           # As a fallback we try to fetch git information from the local git repository
           class LocalGit < Base
             def git_repository_url
-              exec_git_command("git ls-remote --get-url")
+              Utils::Git.exec_git_command("git ls-remote --get-url")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git repository url: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -22,7 +20,7 @@ module Datadog
             end
 
             def git_commit_sha
-              exec_git_command("git rev-parse HEAD")
+              Utils::Git.exec_git_command("git rev-parse HEAD")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git commit SHA: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -31,7 +29,7 @@ module Datadog
             end
 
             def git_branch
-              exec_git_command("git rev-parse --abbrev-ref HEAD")
+              Utils::Git.exec_git_command("git rev-parse --abbrev-ref HEAD")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git branch: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -40,7 +38,7 @@ module Datadog
             end
 
             def git_tag
-              exec_git_command("git tag --points-at HEAD")
+              Utils::Git.exec_git_command("git tag --points-at HEAD")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git tag: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -49,7 +47,7 @@ module Datadog
             end
 
             def git_commit_message
-              exec_git_command("git show -s --format=%s")
+              Utils::Git.exec_git_command("git show -s --format=%s")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git commit message: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -82,7 +80,7 @@ module Datadog
             end
 
             def workspace_path
-              exec_git_command("git rev-parse --show-toplevel")
+              Utils::Git.exec_git_command("git rev-parse --show-toplevel")
             rescue => e
               Datadog.logger.debug(
                 "Unable to read git base directory: #{e.class.name} #{e.message} at #{Array(e.backtrace).first}"
@@ -91,25 +89,6 @@ module Datadog
             end
 
             private
-
-            def exec_git_command(cmd)
-              out, status = Open3.capture2e(cmd)
-
-              raise "Failed to run git command #{cmd}: #{out}" unless status.success?
-
-              # Sometimes Encoding.default_external is somehow set to US-ASCII which breaks
-              # commit messages with UTF-8 characters like emojis
-              # We force output's encoding to be UTF-8 in this case
-              # This is safe to do as UTF-8 is compatible with US-ASCII
-              if Encoding.default_external == Encoding::US_ASCII
-                out = out.force_encoding(Encoding::UTF_8)
-              end
-              out.strip! # There's always a "\n" at the end of the command output
-
-              return nil if out.empty?
-
-              out
-            end
 
             def author
               return @author if defined?(@author)
@@ -127,7 +106,7 @@ module Datadog
 
             def set_git_commit_users
               # Get committer and author information in one command.
-              output = exec_git_command("git show -s --format='%an\t%ae\t%at\t%cn\t%ce\t%ct'")
+              output = Utils::Git.exec_git_command("git show -s --format='%an\t%ae\t%at\t%cn\t%ce\t%ct'")
               unless output
                 Datadog.logger.debug(
                   "Unable to read git commit users: git command output is nil"
